requirements.txt
gradio>=4.0.0
pandas
python-dotenv
huggingface_hub
plotly